+++ Example/Pods/Target Support Files/TraceLog-OSX/TraceLog-OSX-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_TraceLog_OSX : NSObject
-@end
-@implementation PodsDummy_TraceLog_OSX
-@end