+++ Example/Pods/Target Support Files/TraceLog-macOS/TraceLog-macOS-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_TraceLog_macOS : NSObject
+@end
+@implementation PodsDummy_TraceLog_macOS
+@end