+++ Example/Pods/Target Support Files/TraceLog.default-ObjC-OSX/TraceLog.default-ObjC-OSX-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_TraceLog_default_ObjC_OSX : NSObject
-@end
-@implementation PodsDummy_TraceLog_default_ObjC_OSX
-@end