+++ Example/Pods/Target Support Files/TraceLog.default-ObjC-macOS/TraceLog.default-ObjC-macOS-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_TraceLog_default_ObjC_macOS : NSObject
+@end
+@implementation PodsDummy_TraceLog_default_ObjC_macOS
+@end