+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Change Log
 All significant changes to this project will be documented in this file.
 
+## [3.0.0](https://github.com/tonystone/tracelog/tree/3.0.0)
+
+#### Updated
+- Xcode projects to be swift 4.1 compatible.
+
+#### Removed
+- Removed deprecated `TLLogger.configure()`. Use TraceLog.configure in swift instead.
+- Removed deprecated `TLLogger.configureWithEnvironment`.  Use TraceLog.configure in swift instead.
+
+
 ## [2.2.0](https://github.com/tonystone/tracelog/tree/2.2.0)
 
 #### Updated
-- Change build environment requirements to xcode9.2 / Swift 4.0.3 (note: this is ust the build, targets are still the same).
+- Change build environment requirements to xcode9.2 / Swift 4.0.3 (note: this is just the build, targets are still the same).
 
 ## [2.1.0](https://github.com/tonystone/tracelog/tree/2.1.0)